+++ aeron-driver/src/main/c/aeron_retransmit_handler.c
@@ -113,12 +113,12 @@ int aeron_retransmit_handler_on_nak(
             {
                 result = resend(resend_clientd, term_id, term_offset, action->length);
                 action->state = AERON_RETRANSMIT_ACTION_STATE_LINGERING;
-                action->expiry_ns = now_ns + handler->linger_timeout_ns;
+                action->expiry_ns = now_ns + (int64_t)handler->linger_timeout_ns;
             }
             else
             {
                 action->state = AERON_RETRANSMIT_ACTION_STATE_DELAYED;
-                action->expiry_ns = now_ns + handler->delay_timeout_ns;
+                action->expiry_ns = now_ns + (int64_t)handler->delay_timeout_ns;
             }
 
             if (aeron_int64_to_ptr_hash_map_put(&handler->active_retransmits_map, key, action) < 0)
@@ -166,7 +166,7 @@ int aeron_retransmit_handler_process_timeouts(
                 {
                     result = resend(resend_clientd, action->term_id, action->term_offset, action->length);
                     action->state = AERON_RETRANSMIT_ACTION_STATE_LINGERING;
-                    action->expiry_ns = now_ns + handler->linger_timeout_ns;
+                    action->expiry_ns = now_ns + (int64_t)handler->linger_timeout_ns;
                     result++;
                 }
